Webcam/requirements.txt
openai-whisper @ git+https://github.com/openai/whisper.git
sounddevice
soundfile
numpy
firebase-admin
torch
pip install insightface onnxruntime opencv-python firebase-admin flask